techdebt: clean up connection_handler_impl dependency issue (#15815)
Follow up of #15553.
Fixes #15126
