+++ source/common/quic/envoy_quic_dispatcher.h
@@ -18,6 +18,7 @@
 
 #include "envoy/network/listener.h"
 #include "server/connection_handler_impl.h"
+#include "server/active_listener_base.h"
 
 namespace Envoy {
 namespace Quic {
