+++ source/common/quic/envoy_quic_proof_source.h
@@ -3,6 +3,7 @@
 #include "common/quic/envoy_quic_proof_source_base.h"
 #include "common/quic/quic_transport_socket_factory.h"
 
+#include "server/active_listener_base.h"
 #include "server/connection_handler_impl.h"
 
 namespace Envoy {
